Bootstrap the App with the correct .sql files
Bootstraps the app with pre-configured .sql files, as normall done in
the material-calendar-api repo. The startup.sql file is a modified
version of the startup.sql file found at material-calendar-api/sql
and the material_calendar.sql file is the same one found in that
directory.

For best compatibility, the files in this repo should be replaced
with those from the material-calendar-api repo